pyproject.toml
[tool.poetry]
name = "olympiads_project"
version = "0.1.0"
description = "Olympiads project for the best University in the Universe"
authors = ["Lisovskiy <you@example.com>"]

[tool.poetry.dependencies]
python = "^3.9"
fastapi = "^0.65.2"
SQLAlchemy = "^1.4.19"
asyncpg = "^0.23.0"
python-dotenv = "^0.18.0"
psycopg2 = "^2.9.1"
alembic = "^1.6.5"
aiofiles = "^0.7.0"
watchgod = "^0.7"
uvicorn = "^0.15.0"
bcrypt = "^3.2.0"
python-jose = {extras = ["cryptography"], version = "^3.3.0"}
python-multipart = "^0.0.5"

[tool.poetry.dev-dependencies]
mypy = "^0.910"

[build-system]
requires = ["poetry-core>=1.0.0"]
build-backend = "poetry.core.masonry.api"